Refactor folder controller
